requirements.txt
paramiko
CloudFlare
discord.py
python-dotenv